pyproject.toml
[project]
name = "ai-sdr"
version = "0.1.0"
description = "AI SDR using XAI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "duckdb>=1.3.2",
    "email-validator>=2.1.0",
    "fastapi[standard]>=0.116.1",
    "jinja2>=3.1.2",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "reflex>=0.8.7",
    "ruff>=0.12.9",
    "xai-sdk>=1.0.1",
]
